% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covidReff-package.R
\docType{package}
\name{covidReff-package}
\alias{covidReff}
\alias{covidReff-package}
\title{covidReff}
\description{
Simulate Covid based on reproduction and vaccination inputs.
}
\author{
\strong{Maintainer}: Will Mackey \email{wfmackey@gmail.com}

Authors:
\itemize{
  \item Hugh Parsonage
}

}
\keyword{internal}
