% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-efficacy.R
\name{get_vaccine_efficacy}
\alias{get_vaccine_efficacy}
\title{Get vaccine efficacy}
\usage{
get_vaccine_efficacy(
  protection_against,
  vaccine_type,
  vaccine_dose,
  days_since_dosage
)
}
\arguments{
\item{protection_against}{a character vector or element: "poi" (protection against infection) or "poh" (protection against hospitalisation)}

\item{vaccine_type}{a character element "pf" (Pfizer) or "az" (Astrazeneca)}

\item{vaccine_dose}{a numeric element 1 for first dose or 2 for second dose}

\item{days_since_dosage}{a numeric element 1 for first dose or 2 for second dose}
}
\description{
Return an efficacy value based on protection (infection or hospitalisation), vaccine type, dose and days since dosage.
}
