% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate-covid.R
\name{simulate_covid}
\alias{simulate_covid}
\title{Simulate COVID}
\arguments{
\item{r0}{base reproduction number. Defaults to 3 plus delta variant increases by 50 per cent: https://assets.publishing.service.gov.uk/government/uploads/system/uploads/attachment_data/file/993232/S1272_LSHTM_Modelling_Paper_B.1.617.2.pdf}

\item{serial_interval}{5 days, from https://www.medrxiv.org/content/10.1101/2021.06.04.21258205v1.full.pdf}

\item{vaccination_levels}{vaccination_levels by age; provided in a named vector.}

\item{uniform_vaccination_rate}{defaults  to NULL.}

\item{weekly_vaccinations}{defaults to  0.005. The additional proportion of the population vaccinated each week}

\item{p_max_vaccinated}{defaults to 0.90.}

\item{vac_infection_rate}{defaults to 0.2.}

\item{vac_transmission_rate}{defaults to 0.5.}

\item{vac_hospitalisation_reduction}{defaults to 0.95. The hospitalisation reduction when vaccinated GIVEN infection}

\item{vac_death_reduction}{defaults to 0.99. The death reduction when vaccinated GIVEN infection}

\item{hospitalisation_per_death}{defaults to 20.}

\item{max_hospitalisation_rate}{defaults to 0.95.}

\item{population_scale_factor}{defaults to 10, where values of 1 implies 26m  population; 10=2.6m, 100=260k population, etc}

\item{n_start_infected}{defaults to 100 people infected at day 0.}

\item{p_max_infected}{defaults to 0.8. The proportion who CAN get infected if it spreads; kinda like herd immunity level}

\item{n_iterations}{defaults to  3.}

\item{simulations}{defaults to  1.}

\item{scenario}{defaults to 1.}

\item{return_iterations}{defaults to TRUE}

\item{return_population}{defaults to FALSE}
}
\value{
A \code{tibble} object.
}
\description{
simulate covid for a given reproduction number, level of
vaccinations in a population, and other epidemiological params.
}
