% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vaccine-efficacy-documentation.R
\docType{data}
\name{ved_list}
\alias{ved_list}
\title{Vaccine efficacy data by day}
\format{
A list containing eight named numeric vectors of length 600:
\describe{
\item{\code{ved_poi_pf1}}{Vaccine protection against infection after a single dose of Pfizer from day 1 to 600 after vaccination.}
\item{\code{ved_poi_pf2}}{Vaccine protection against infection after a double dose of Pfizer from day 1 to 600 after vaccination.}
\item{\code{ved_poi_az1}}{Vaccine protection against infection after a single dose of AstraZeneca from day 1 to 600 after vaccination.}
\item{\code{ved_poi_az2}}{Vaccine protection against infection after a double dose of AstraZeneca from day 1 to 600 after vaccination.}
\item{\code{ved_poh_pf1}}{Vaccine protection against hospitalisation after a single dose of Pfizer from day 1 to 600 after vaccination.}
\item{\code{ved_poh_pf2}}{Vaccine protection against hospitalisation after a double dose of Pfizer from day 1 to 600 after vaccination.}
\item{\code{ved_poh_az1}}{Vaccine protection against hospitalisation after a single dose of AstraZeneca from day 1 to 600 after vaccination.}
\item{\code{ved_poh_az2}}{Vaccine protection against hospitalisation after a double dose of AstraZeneca from day 1 to 600 after vaccination.}
}
}
\usage{
ved_list
}
\description{
Linear interpolation of vaccine efficacy data from day 1 to 600 since dosage. See \code{?vaccine_efficacy}
}
\keyword{datasets}
