% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-vaccination-rates.R
\name{get_population_rate}
\alias{get_population_rate}
\title{Get population rate based on an age-specific input}
\arguments{
\item{age_rate}{a named vector with levels for 'under12', 'under40', 'under60',
'under80', 'over80'; eg: c(under12 = 0.00, under40 = 0.70, under60 = 0.90, under80 = 0.95, over80  = 0.95)}
}
\description{
determine the Australian population level rate for age-specific rates
}
